+++ src/main/java/com/carpool/car_pool/repositories/RideRequestRepository.java
@@ -1,0 +1,9 @@
+package com.carpool.car_pool.repositories;
+
+import com.carpool.car_pool.repositories.entities.RideRequestsEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface RideRequestRepository extends JpaRepository<RideRequestsEntity, Long> {
+}
