+++ src/main/java/com/carpool/car_pool/repositories/UserRepository.java
@@ -1,0 +1,9 @@
+package com.carpool.car_pool.repositories;
+
+import com.carpool.car_pool.repositories.entities.UserEntity;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface UserRepository extends JpaRepository<UserEntity, Long> {
+}
