+++ src/main/java/com/carpool/car_pool/repositories/entities/RideOfferEntity.java
@@ -1,11 +1,11 @@
-package com.carpool.car_pool.entities;
+package com.carpool.car_pool.repositories.entities;
 
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.Id;
 import jakarta.persistence.ManyToMany;
-import jakarta.persistence.ManyToOne;
+import jakarta.persistence.OneToMany;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
@@ -20,16 +20,17 @@ import java.util.List;
 @NoArgsConstructor
 @AllArgsConstructor
 @Entity
-public class RideRequestsEntity {
-    
+public class RideOfferEntity {
+
     @Id
     @GeneratedValue
     private Long id;
 
-    @ManyToOne
-    private RideOfferEntity rideOffer;
 
-    @ManyToMany(mappedBy = "rideRequests")
+    @ManyToMany(mappedBy = "rideOffers")
     @JsonIgnore
-    private List<UserEntity> user;
+    private List<UserEntity> users;
+
+    @OneToMany(mappedBy = "rideOffer")
+    private List<RideRequestsEntity> rideRequests;
 }
