+++ src/main/java/com/carpool/car_pool/repositories/entities/UserEntity.java
@@ -1,4 +1,4 @@
-package com.carpool.car_pool.entities;
+package com.carpool.car_pool.repositories.entities;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
